Move a chain contract interface to chaindriver
