+++ node/action/parameter.go
@@ -7,7 +7,7 @@ package action
 
 import (
 	"github.com/Oneledger/protocol/node/chains/bitcoin"
-	"github.com/Oneledger/protocol/node/chains/common"
+	"github.com/Oneledger/protocol/node/chains/driver"
 	"github.com/Oneledger/protocol/node/chains/ethereum"
 	"github.com/Oneledger/protocol/node/data"
 	"github.com/Oneledger/protocol/node/id"
@@ -189,14 +189,14 @@ func GetByte32(value FunctionValue) [32]byte {
 //	return INVALID
 //}
 //
-func GetContract(value FunctionValue) common.Contract {
+func GetContract(value FunctionValue) chaindriver.Contract {
 	switch value.(type) {
 	case bitcoin.HTLContract:
 		return value.(*bitcoin.HTLContract)
 	case ethereum.HTLContract:
 		return value.(*ethereum.HTLContract)
-	case common.Contract:
-		return value.(common.Contract)
+	case chaindriver.Contract:
+		return value.(chaindriver.Contract)
 	default:
 		log.Fatal("Bad Type Cast in Function Parameter", "value", value)
 	}
