--- node/chains/common/driver.go
@@ -1,12 +1,0 @@
-package common
-
-import (
-	"github.com/Oneledger/protocol/node/data"
-)
-
-type Contract interface {
-	Chain() data.ChainType
-	ToBytes() []byte
-	ToKey() []byte
-	FromBytes([]byte)
-}
